pyproject.toml
[tool.poetry]
name = "editor-demo"
version = "0.1.0"
description = ""
authors = ["Daehan Kim <daehankim@ibigtree.kr>"]

[tool.poetry.dependencies]
python = "~3.9"
python-dotenv = "^0.19.0"
Django = "^3.2.7"
gunicorn = {version = "^20.1.0", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
sphinx-rtd-theme = "^1.0.0"
sphinxcontrib-apidoc = "^0.3.0"
flake8 = "^3.9.2"
autopep8 = "^1.5.7"
Sphinx = "^4.2.0"
myst-parser = "^0.15.2"
sphinxcontrib-plantuml = "^0.21"
flake8-import-order = "^0.18.1"
flake8-mypy = "^17.8.0"

[tool.poetry.extras]
wsgi = ["gunicorn"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.masonry.api"
